requirements.txt
Flask
Flask-Bcrypt
psycopg2-binary
gunicorn